This commit resolves a `decimal.InvalidOperation` error that occurred during magnitude filtering and ensures that dwarf planet magnitudes are calculated and handled correctly.
The changes include:
- Modifying the magnitude filtering logic in `get_visible` to explicitly cast the `Magnitude` column to `float` before comparison. This prevents errors when the column contains mixed data types like `Decimal`.
- Implementing a dynamic magnitude calculation for dwarf planets using `ephem.EllipticalBody` and orbital elements from the MPCORB database.
- Centralizing the minor planet name resolution logic in `apts/utils/planetary.py` to handle both simple and full designations robustly.
- Ensuring that the minor planets DataFrame is eagerly loaded in the `SolarObjects` class to be available for magnitude calculations.

+++ apts/objects/solar_objects.py
@@ -284,7 +284,7 @@ class SolarObjects(Objects):
                             lambda x: x.magnitude if hasattr(x, "magnitude") else x
                         ),
                         errors="coerce",
-                    )
+                    ).astype("float")
                     < conditions.max_object_magnitude
                 )
             )
